Use sudo-prompt to re-run tsrelay in Linux (#64)
This (draft) PR prompts the user to enter their password for Linux when
tsrelay is needed to run with `sudo`.

Note that for now it assumes Linux already has `pkexec` installed. 
Currently, it keeps the same nonce and port so that we don't need to let
the client reconfigure its attributes.