Rename view and viewItem contexts for consistency with tree classes
I was finding it difficult to remember the hierarchy of the TreeView, as
the naming has not been consistent over time. This attempts to create
a naming convention for the TreeItem classes and their contextValue.

I also removed "tailscale-" from the view ID's as they need only be
unique within our extension and the larger names made the when clauses
more unruly.

Signed-off-by: Tyler Smalley <tyler@tailscale.com>