File Explorer: Fix issue with spaces in filenames
We were escaping spaces while using that value directly in SFTP calls. This caused the escape character (`\`) to be included in the file/directory name. Additionally, if a space was included it would not be able to be deleted as it would issue the delete call to include the escape character.

Closes #231

Signed-off-by: Tyler Smalley <tyler@tailscale.com>